Changed node positions to double type. Improved uniform number generator.
git-svn-id: https://subversion.assembla.com/svn/elec476/trunk/src@36 4db04919-9836-40c1-aee8-a115cec00ef4

+++ contextAwareRouting/CentralServer.java
@@ -40,7 +40,7 @@ public class CentralServer {
 		}
 	}
 
-	private boolean inRange(int x1, int y1, int x2, int y2){
+	private boolean inRange(double x1, double y1, double x2, double y2){
 		double distance = Math.sqrt(Math.pow(x1-x2, 2) + Math.pow(y1-y2, 2));
 		if (distance <= Mainline.T)
 			return true;
